+++ android/assets/src/com/unciv/logic/battle/BattleDamage.kt
@@ -237,7 +237,7 @@ object BattleDamage {
                 modifiers["Tile"] = (tileDefenceBonus * 100).toInt()
 
 
-            if (defender.unit.isFortified() || defender.unit.isGuarding())
+            if (defender.unit.isFortified() || defender.unit.isGuarding() || defender.unit.isAlerted())
                 modifiers["Fortification"] = BattleConstants.FORTIFICATION_BONUS * defender.unit.getFortificationTurns()
         }