+++ android/assets/src/com/unciv/logic/map/mapunit/UnitUpgradeManager.kt
@@ -109,7 +109,7 @@ class UnitUpgradeManager(val unit: MapUnit) {
         if (!isFree) civ.addGold(-(goldCostOfUpgrade ?: getCostOfUpgrade(upgradedUnit)))
         newUnit.currentMovement = 0f
         // wake up if lost ability to fortify
-        if (newUnit.isFortified() && !newUnit.canFortify(ignoreAlreadyFortified = true))
+        if ((newUnit.isFortified() || newUnit.isAlerted()) && !newUnit.canFortify(ignoreAlreadyFortified = true))
             newUnit.action = null
         // wake up from Guarding if can't Withdraw
         if (newUnit.isGuarding() && !newUnit.hasUnique(UniqueType.WithdrawsBeforeMeleeCombat))