+++ core/src/com/unciv/ui/components/widgets/UnitIconGroup.kt
@@ -136,7 +136,7 @@ class UnitIconGroup(val unit: MapUnit, val size: Float) : NonTransformGroup() {
     private fun getBackgroundDrawableForUnit(): TextureRegionDrawable {
         return when {
             unit.isEmbarked() -> ImageGetter.getDrawable("UnitFlagIcons/UnitFlagEmbark")
-            unit.isFortified() -> ImageGetter.getDrawable("UnitFlagIcons/UnitFlagFortify")
+            unit.isFortified() || unit.isAlerted() -> ImageGetter.getDrawable("UnitFlagIcons/UnitFlagFortify")
             unit.isGuarding() -> ImageGetter.getDrawable("UnitFlagIcons/UnitFlagFortify")
             unit.isCivilian() -> ImageGetter.getDrawable("UnitFlagIcons/UnitFlagCivilian")
             else -> ImageGetter.getDrawable("UnitFlagIcons/UnitFlag")
@@ -146,7 +146,7 @@ class UnitIconGroup(val unit: MapUnit, val size: Float) : NonTransformGroup() {
     private fun getBackgroundInnerDrawableForUnit(): TextureRegionDrawable? {
         return when {
             unit.isEmbarked() -> ImageGetter.getDrawableOrNull("UnitFlagIcons/UnitFlagEmbarkInner")
-            unit.isFortified() -> ImageGetter.getDrawableOrNull("UnitFlagIcons/UnitFlagFortifyInner")
+            unit.isFortified() || unit.isAlerted() -> ImageGetter.getDrawableOrNull("UnitFlagIcons/UnitFlagFortifyInner")
             unit.isGuarding() -> ImageGetter.getDrawableOrNull("UnitFlagIcons/UnitFlagFortifyInner")
             unit.isCivilian() -> ImageGetter.getDrawableOrNull("UnitFlagIcons/UnitFlagCivilianInner")
             else -> ImageGetter.getDrawableOrNull("UnitFlagIcons/UnitFlagInner")
@@ -157,7 +157,7 @@ class UnitIconGroup(val unit: MapUnit, val size: Float) : NonTransformGroup() {
 
         val filename = when {
             unit.isEmbarked() -> "UnitFlagIcons/UnitFlagMaskEmbark"
-            unit.isFortified() -> "UnitFlagIcons/UnitFlagMaskFortify"
+            unit.isFortified() || unit.isAlerted() -> "UnitFlagIcons/UnitFlagMaskFortify"
             unit.isGuarding() -> "UnitFlagIcons/UnitFlagMaskFortify"
             unit.isCivilian() -> "UnitFlagIcons/UnitFlagMaskCivilian"
             else -> "UnitFlagIcons/UnitFlagMask"
@@ -171,7 +171,7 @@ class UnitIconGroup(val unit: MapUnit, val size: Float) : NonTransformGroup() {
     private fun getBackgroundSelectionForUnit(): Image {
         return when {
             unit.isEmbarked() -> ImageGetter.getImage("UnitFlagIcons/UnitFlagSelectionEmbark")
-            unit.isFortified() -> ImageGetter.getImage("UnitFlagIcons/UnitFlagSelectionFortify")
+            unit.isFortified() || unit.isAlerted() -> ImageGetter.getImage("UnitFlagIcons/UnitFlagSelectionFortify")
             unit.isGuarding() -> ImageGetter.getImage("UnitFlagIcons/UnitFlagSelectionFortify")
             unit.isCivilian() -> ImageGetter.getImage("UnitFlagIcons/UnitFlagSelectionCivilian")
             else -> ImageGetter.getImage("UnitFlagIcons/UnitFlagSelection")